requirements.txt
Django
mysql-connector-python
openpyxl
Pillow
PyMySQL
pytz
djangorestframework
django-filter